pyproject.toml
[project]
name = "psql-connection-load"
version = "0.1.0"
description = ""
authors = [
    {name = "Alexandre J. Martins",email = "alexandre@sigalei.com.br"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "psycopg (>=3.3.1,<4.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
